Raycast from camera to mouse point. Basic rotation of camera around skull upon right click.
There is a raycast cast from the camera to wherever the mouse is on the screen. When this raycast hits a collider on the skull, it changes a UI element to display the colliders name. In the future we should have more precise mesh colliders. The mouse position is an event stored in InputController. There is also an event in that file for when the right mouse button is held. Currently holding down the mouse button allows the camera to rotate around the skull in a horizontal motion. The camera has been made to rotate with the skull as its pivot point.